Progression of AST
Header file (ast.mli):
- nearly finished expressions (type expr) but as can be seen from the *INSIGHTFUL* comments everything with **ttype** was
left blank-unit- on purpose since the symbol table has to be examined and created(thus, we need to study :) firstly)
- statements are finished
- declaration is in immediate need of debugging and fixing since:
-- how can a record field have a type of tuple, i've written DeclList*StmtList but it causes a syntax error
-- parameter type needs to be further examined, right now it is declared as **string list** but maybe ***byref** should be somehow declared??
or else symbol table will take care of that
-- *declarator* is either a string or an array being indexed
- last point brings us to another problem with *const_expr*
the discussion revolved around whether we should somehow evaluate the expression's value to an integer and then index the array
but that means that we cant write this **a[a[]]** right? so how should we tackle this?
P.S 1:  maybe talk about when a declarator is invoked and what a constant expression actually means(too generic, not a great help)
P.S 2:  syntax error so *make* fails but im going to sleep